\name{zhouyahomework2-package}
\alias{zhouyahomework2-package}
\alias{zhouyahomework2}
\docType{package}
\title{
\packageTitle{zhouyahomework2}
}
\description{
\packageDescription{zhouyahomework2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{zhouyahomework2}
\packageIndices{zhouyahomework2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{zhouyahomework2}

Maintainer: \packageMaintainer{zhouyahomework2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
